Added validation for phone number

+++ db/migrate/20181109050728_create_users.rb
@@ -2,7 +2,7 @@ class CreateUsers < ActiveRecord::Migration[5.2]
   def change
     create_table :users do |t|
       t.string :name
-      t.decimal :phone_num
+      t.integer :phone_num
       t.timestamps null: false
     end
   end
